rvc_engine.py
import os
import logging
import torch
import soundfile as sf
import subprocess

# Try imports, if fail, use mock
try:
    from faiss import read_index
    from fairseq import checkpoint_utils
    # We would need the full RVC codebase here, but for MVP we might wrap a CLI or simple script
    # RVC is complex. Let's use a simplified approach or just Mock for now if dependencies fail.
    HAS_RVC = False # Set to true once installed
except ImportError:
    HAS_RVC = False

logger = logging.getLogger("soniqstem-api")

def infer_audio(audio_path, model_name, pitch_shift):
    """
    RVC Inference Pipeline.
    """
    logger.info(f"RVC Inference: {audio_path} -> {model_name}, Pitch: {pitch_shift}")
    
    output_path = audio_path.replace(".wav", "").replace(".mp3", "") + f"_{model_name}_cloned.wav"

    if HAS_RVC:
        # Conceptual RVC Logic:
        # 1. Load Hubert
        # 2. Load Model & Index
        # 3. Process
        pass
    else:
        logger.warning("RVC libraries not found. Using Mock Conversion (Pitch Shift Only).")
        # Fallback: Just simple pitch shift to simulate "changing voice"
        import librosa
        y, sr = librosa.load(audio_path, sr=None)
        
        # Simple heuristic: 
        # Male to Female (+12) -> Pitch shift up
        # Female to Male (-12) -> Pitch shift down
        steps = float(pitch_shift)
        if steps != 0:
            y_shifted = librosa.effects.pitch_shift(y, sr=sr, n_steps=steps)
        else:
            y_shifted = y
            
        sf.write(output_path, y_shifted, sr)
        
    return output_path
