requirements.txt
demucs
torch
torchaudio
librosa
numpy
soundfile
transformers
scipy
accelerate
matchering
fairseq
faiss-cpu
torchcrepe
tensorboardX
praat-parselmouth
